requirements.txt
cryptography
Django
django-bootstrap5
django-filter
djangorestframework
drf-spectacular
fontawesomefree
markdown
psycopg2-binary
pyjwt
python-dateutil
requests
uwsgi